// src/api/models/storeListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface StoreListItem {
  /** ID */
  id: string;
  /** 이름 */
  name: string;
  /** 전화번호 */
  phoneNumber: string;
  /** 주소 */
  address: string;
  /**
   * 이용권 가격
   * @nullable
   */
  passPrice?: string | null;
  /** 생성일 */
  createdAt: string;
  /** 삭제 여부 */
  isDeleted: boolean;
}